debugging image output

+++ data_utils.py
@@ -144,7 +144,7 @@ def get_open_images(id_data, class_limit):
 
     for image_id, group in image_data:
         classes = group['RealClass'].as_matrix()
-        obj_vis = np.array([1 if name in classes else 0 for name in OFFICIAL_CLASS_LIST])
+        obj_vis = np.array([1 if name in classes else 0 for name in OFFICIAL_CLASS_LIST], dtype=np.uint8)
         image_url = list(group['OriginalURL'])[0]
         image_file = output_image_dir + '/{}.jpg'.format(image_id)
         image_bytes = BytesIO(requests.get(image_url).content)
@@ -167,7 +167,7 @@ def get_coco_images(id_data, class_limit):
 
     for image_id, group in id_data:
         classes = group['RealClass'].as_matrix()
-        obj_vis = np.array([1 if name in classes else 0 for name in OFFICIAL_CLASS_LIST])
+        obj_vis = np.array([1 if name in classes else 0 for name in OFFICIAL_CLASS_LIST], dtype=np.uint8)
         id_str = pad_img_num(image_id, COCO_ID_LENGTH)
         image = misc.imread(coco_image_file.format(id_str)).astype(np.uint8)
         torch.save({'frame':image, 'obj_vis':obj_vis}, output_image_file.format(id_str))
